requirements.txt

pandas==1.3.3
numpy==1.21.2
tweepy==4.4.0
# linkedin-api==2.0.0
facebook-sdk==3.1.0
schedule==1.1.0
scikit-learn==0.24.2
transformers==4.42.4
torch==2.3.1
spacy==3.1.3
textstat==0.7.2
gensim==4.1.2
pyLDAvis==3.3.1
openai==0.27.0
scipy==1.7.1
PyQt5==5.15.4
pyinstaller==4.5.1
